последняя цифра трехзначного числа

+++ task№6/Program.cs
@@ -1,2 +1,2 @@
-﻿double numberA = 764;
+﻿int numberA = 764;
 Console.WriteLine(numberA%10);
